requirements.txt
wheel
numpy
textblob
nltk
libhxl

